[TSK] Refatorando o Projeto [REF] Ajeitando o projeto em diretorios e colocando textos em constantes
