+++ lib/components/editor.dart
@@ -1,0 +1,28 @@
+import 'package:flutter/cupertino.dart';
+import 'package:flutter/material.dart';
+
+class Editor extends StatelessWidget {
+  final TextEditingController _controller;
+  final String _rotulo;
+  final String _dica;
+  final IconData icone;
+
+  Editor(this._controller, this._rotulo, this._dica, {this.icone});
+
+  @override
+  Widget build(BuildContext context) {
+    return Padding(
+      padding: const EdgeInsets.all(16.0),
+      child: TextField(
+        controller: _controller,
+        keyboardType: TextInputType.number,
+        style: TextStyle(fontSize: 24.0),
+        decoration: InputDecoration(
+          icon: icone != null ? Icon(icone) : null,
+          labelText: _rotulo,
+          hintText: _dica,
+        ),
+      ),
+    );
+  }
+}
