+++ lib/main.dart
@@ -1,3 +1,4 @@
+import 'package:bytebank/screens/transferencia/lista.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 
@@ -20,153 +21,3 @@ class ByteBank extends StatelessWidget {
   }
 }
 
-class FormularioTransferencia extends StatefulWidget {
-
-  @override
-  State<StatefulWidget> createState() {
-    return FormularioTransferenciaState();
-  }
-}
-
-class FormularioTransferenciaState extends State<FormularioTransferencia> {
-  final TextEditingController _controladorCampoNumeroConta =
-  TextEditingController();
-  final TextEditingController _controladorCampoValor = TextEditingController();
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: Text('Crinado Transferencias'),
-      ),
-      body: SingleChildScrollView(
-        child: Column(
-          children: <Widget>[
-            Editor(
-              _controladorCampoNumeroConta,
-              'Numero da conta',
-              '0000',
-            ),
-            Editor(
-              _controladorCampoValor,
-              'Valor',
-              '000.0',
-              icone: Icons.monetization_on,
-            ),
-            RaisedButton(
-              child: Text('Confirmar'),
-              onPressed: () => _criaTransferencia(context),
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-
-  void _criaTransferencia(context) {
-    final int numeroConta =
-        int.tryParse(_controladorCampoNumeroConta.text);
-    final double valor = double.tryParse(_controladorCampoValor.text);
-
-    if (numeroConta != null && valor != null) {
-      final transferencia = Transferencia(valor, numeroConta);
-      Navigator.pop(context, transferencia);
-    }
-  }
-}
-
-class Editor extends StatelessWidget {
-  final TextEditingController _controller;
-  final String _rotulo;
-  final String _dica;
-  final IconData icone;
-
-  Editor(this._controller, this._rotulo, this._dica, {this.icone});
-
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-      padding: const EdgeInsets.all(16.0),
-      child: TextField(
-        controller: _controller,
-        keyboardType: TextInputType.number,
-        style: TextStyle(fontSize: 24.0),
-        decoration: InputDecoration(
-          icon: icone != null ? Icon(icone) : null,
-          labelText: _rotulo,
-          hintText: _dica,
-        ),
-      ),
-    );
-  }
-}
-
-class ListaTransferencia extends StatefulWidget {
-  final List<Transferencia> _listaTransferencia = List();
-
-  @override
-  State<StatefulWidget> createState() {
-    return ListaTransferenciaState();
-  }
-}
-
-class ListaTransferenciaState extends State<ListaTransferencia> {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(title: Text('Transferencias')),
-      floatingActionButton: FloatingActionButton(
-        child: Icon(Icons.add),
-        onPressed: () {
-          final Future<Transferencia> future = Navigator.push(
-              context,
-              MaterialPageRoute(
-                  builder: (BuildContext context) =>
-                      FormularioTransferencia()));
-          future.then((item) {
-            if (item != null) setState(() {
-              widget._listaTransferencia.add(item);
-            });
-          });
-        },
-      ),
-      body: ListView.builder(
-        itemCount: widget._listaTransferencia.length,
-        itemBuilder: (context, index) {
-          final transferencia = widget._listaTransferencia[index];
-          return ItemTransferencia(transferencia);
-        },
-      ),
-    );
-  }
-}
-
-class ItemTransferencia extends StatelessWidget {
-  final Transferencia _transferencia;
-
-  ItemTransferencia(this._transferencia);
-
-  @override
-  Widget build(BuildContext context) {
-    // TODO: implement build
-    return Card(
-      child: ListTile(
-        leading: Icon(Icons.monetization_on),
-        title: Text(_transferencia.valor.toString()),
-        subtitle: Text(_transferencia.numeroConta.toString()),
-      ),
-    );
-  }
-}
-
-class Transferencia {
-  final double valor;
-  final int numeroConta;
-
-  Transferencia(this.valor, this.numeroConta);
-
-  @override
-  String toString() {
-    return 'Transferencia{valor: $valor, numeroConta: $numeroConta}';
-  }
-}
